Refactored differentiation classes into separate files.
