+++ sindy/differentiation/__init__.py
@@ -1,1 +1,3 @@
-from .differentiation_methods import *
+from .base import BaseDifferentiation
+from .finite_difference import FiniteDifference
+from .smoothed_finite_difference import SmoothedFiniteDifference
