+++ sindy/differentiation/finite_difference.py
@@ -1,47 +1,6 @@
-import abc
 import numpy as np
-from scipy.signal import savgol_filter
 
-from sindy.utils.base import validate_input
-
-
-class BaseDifferentiation:
-    """
-    Base class for differentiation methods.
-
-    Simply forces differentiation methods to implement a
-    _differentiate function.
-    """
-    def __init__(self):
-        pass
-
-    # Force subclasses to implement this
-    @abc.abstractmethod
-    def _differentiate(self, x, t):
-        """
-        Numerically differentiate data.
-
-        Parameters
-        ----------
-        x: array-like, shape (n_samples, n_input_features)
-            Data to be differentiated. Rows of x should correspond to the same
-            point in time.
-
-        t: float or numpy array of shape [n_samples]
-            If t is a float, it is interpreted as the timestep between
-            samples in x.
-            If t is a numpy array, it specifies the times corresponding
-            to the rows of x. That is, t[i] should be the time at which
-            the measurements x[i, :] were taken.
-            The points in t are assumed to be increasing.
-
-        Returns
-        -------
-        x_dot: array-like, shape (n_samples, n_input_features)
-            Numerical time derivative of x. Entries where derivatives were
-            not computed will have the value np.nan.
-        """
-        raise NotImplementedError
+from sindy.differentiation import BaseDifferentiation
 
 
 class FiniteDifference(BaseDifferentiation):
@@ -85,10 +44,6 @@ class FiniteDifference(BaseDifferentiation):
             return self._forward_difference(x, t)
         else:
             return self._centered_difference(x, t)
-
-    def __call__(self, x, t=1):
-        x = validate_input(x)
-        return self._differentiate(x, t)
 
     def _forward_difference(self, x, t=1):
         """
@@ -188,47 +143,3 @@ class FiniteDifference(BaseDifferentiation):
                     / (t_diff[-1] / 2)
                 )
             return x_dot
-
-
-class SmoothedFiniteDifference(FiniteDifference):
-    """
-    Perform differentiation by smoothing input data then applying a finite
-    difference method.
-
-    Parameters
-    ----------
-    smoother: function, optional (default savgol_filter)
-        Function to perform smoothing. Must be compatible with the
-        following call signature:
-        x_smoothed = smoother(x, **smoother_kws)
-
-    smoother_kws: dict, optional (default {})
-        Arguments passed to smoother when it is invoked.
-
-    **kwargs: kwargs
-        Addtional parameters passed to the FiniteDifference __init__
-        function.
-    """
-    def __init__(
-        self,
-        smoother=savgol_filter,
-        smoother_kws={},
-        **kwargs
-    ):
-        super(SmoothedFiniteDifference, self).__init__(**kwargs)
-        self.smoother = smoother
-        self.smoother_kws = smoother_kws
-
-        if smoother is savgol_filter:
-            if 'window_length' not in smoother_kws:
-                self.smoother_kws['window_length'] = 11
-            if 'polyorder' not in smoother_kws:
-                self.smoother_kws['polyorder'] = 3
-            self.smoother_kws['axis'] = 0
-
-    def _differentiate(self, x, t):
-        """
-        Apply finite difference method after smoothing.
-        """
-        x = self.smoother(x, **self.smoother_kws)
-        return super(SmoothedFiniteDifference, self)._differentiate(x, t)
